LLM/requirements.txt
# LLM-based Sophisticated Fake News Detection System Requirements

# Core LLM and AI
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0
bitsandbytes>=0.39.0
sentencepiece>=0.1.99
sentence-transformers>=2.2.0

# Web Scraping and HTTP
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.10.0
aiohttp>=3.8.0
httpx>=0.24.0
scrapy>=2.9.0
newspaper3k>=0.2.8

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
jsonschema>=4.17.0

# Text Processing and NLP
spacy>=3.6.0
nltk>=3.8.0
langdetect>=1.0.9
textstat>=0.7.3
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0

# News APIs and Social Media
newsapi-python>=0.2.6
praw>=7.7.0  # Reddit API
tweepy>=4.14.0  # Twitter API

# Advanced Analysis
scikit-learn>=1.3.0
scipy>=1.10.0
networkx>=3.1.0

# Utilities
tqdm>=4.65.0
colorama>=0.4.6
python-dateutil>=2.8.2
pydantic>=2.0.0
asyncio-throttle>=1.0.2

# Performance and Caching
redis>=4.5.0
diskcache>=5.6.0
faiss-cpu>=1.7.4  # For vector similarity search

# Monitoring and Logging
structlog>=23.1.0
rich>=13.4.0
